requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
requests==2.32.5
python-dotenv==1.1.1
deep-translator==1.11.4
python-multipart==0.0.20
